feat : VoteResponseDto에 총투표 명수 파라미터 추가 및 투표 재시작 구현

+++ src/main/java/net/causw/application/dto/vote/VoteOptionResponseDto.java
@@ -13,6 +13,6 @@ import java.util.List;
 public class VoteOptionResponseDto {
     private String id;
     private String optionName;
-    private Long voteCount;
+    private Integer voteCount;
     private List<UserResponseDto> voteUsers;
 }
